gen_ai_finops/pyproject.toml
[project]
name = "gen-ai-finops"
version = "0.1.0"
description = "The 'Kubernetes' of AI Costs - Automated Token Optimization & Multi-Cloud Pricing Intelligence"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "GenAIFinOps Team"}
]
keywords = ["ai", "llm", "cost-optimization", "pricing", "finops", "rag", "chromadb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "chromadb>=0.4.22",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "litellm>=1.17.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "tiktoken>=0.5.2",
    "lxml>=5.1.0",
    "html5lib>=1.1",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
